+++ CHANGELOG.md
@@ -18,6 +18,7 @@
   * flutter pub upgrade major-versions.
   * Including reference to THFile in THElement.
   * Renaming MPBoundingBox to MPBoundingBoxMixin.
+  * Implementing subtypeLineSegmentMPIDsByLineSegmentIndex map in THLine to keep track of which line segments have subtype options.
 
 ## 0.2.26 - 2025-12-11 - The São Paulo release
 * New features: